[AIX] PGO codegen changes for function-sections.
The existing PGO implementation creates a dependence from the profiling
data symbol to the function descriptors for every function defined
within the compilation unit. This forces the binder to include either all
or no functions from an object file when linking a profile-generate
enabled object file, which can break links that require the aggressive
"include only whats used" link semantics that XCOFF provides. To remedy
this we break up the profiling data symbol into per-function CSECTs by
using the rename directive to emit the profd symbol for a single function
to a unique CSECT, then rename it to the common name afterwards. We also
have to split the counters into per-function CSECTs as they need to have a
.ref relocation from the  counter section which is referenced by the
functions, to the profiling-data symbol which is otherwise unreferenced.